S08P12A807-280 #bugfix Feature:back-board: board 게시물에 빠진 속성 추가

See merge request s08-webmobile2-sub2/S08P12A807!125

+++ togetdog/src/main/java/com/ssafy/togetdog/board/model/repository/LikeRepository.java
@@ -10,7 +10,7 @@ import com.ssafy.togetdog.user.model.entity.User;
 
 public interface LikeRepository extends JpaRepository<LikePost, Long> {
 
-	Long countByBoard(Board board);
+	int countByBoard(Board board);
 
 	Optional<LikePost> findByBoardAndUser(Board board, User user);
